mythfilerecorder: rename commandlineparser.* to mythfilerecorder_…
